moved removeHtmlTags and isValidMbidFormat utility funcs to shared module Utils.kt
