+++ app/src/main/java/org/listenbrainz/android/ui/screens/search/BaseSearchScreen.kt
@@ -64,7 +64,7 @@ import org.listenbrainz.android.ui.components.TitleAndSubtitle
 import org.listenbrainz.android.ui.theme.ListenBrainzTheme
 import org.listenbrainz.android.util.Utils.formatDurationSeconds
 import org.listenbrainz.shared.util.Utils.getCoverArtUrl
-import org.listenbrainz.android.util.Utils.removeHtmlTags
+import org.listenbrainz.shared.util.Utils.removeHtmlTags
 import org.listenbrainz.android.viewmodel.SearchViewModel
 import org.listenbrainz.shared.model.feed.FeedListenArtist
 import kotlinx.datetime.LocalDate
